add custom hook with logic of modal show
